admin dashboard now displays all the users
