--- Software_Engineering_Project/Models/RoomsController.cs
@@ -1,126 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.Entity;
-using System.Linq;
-using System.Net;
-using System.Web;
-using System.Web.Mvc;
-
-namespace Software_Engineering_Project.Models
-{
-    public class RoomsController : Controller
-    {
-        private RoomsContext db = new RoomsContext();
-
-        // GET: Rooms
-        public ActionResult Index()
-        {
-            return View(db.Rooms.ToList());
-        }
-
-        // GET: Rooms/Details/5
-        public ActionResult Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Rooms rooms = db.Rooms.Find(id);
-            if (rooms == null)
-            {
-                return HttpNotFound();
-            }
-            return View(rooms);
-        }
-
-        // GET: Rooms/Create
-        public ActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: Rooms/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "id,descriptorID,capacity")] Rooms rooms)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Rooms.Add(rooms);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-
-            return View(rooms);
-        }
-
-        // GET: Rooms/Edit/5
-        public ActionResult Edit(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Rooms rooms = db.Rooms.Find(id);
-            if (rooms == null)
-            {
-                return HttpNotFound();
-            }
-            return View(rooms);
-        }
-
-        // POST: Rooms/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "id,descriptorID,capacity")] Rooms rooms)
-        {
-            if (ModelState.IsValid)
-            {
-                db.Entry(rooms).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
-            return View(rooms);
-        }
-
-        // GET: Rooms/Delete/5
-        public ActionResult Delete(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Rooms rooms = db.Rooms.Find(id);
-            if (rooms == null)
-            {
-                return HttpNotFound();
-            }
-            return View(rooms);
-        }
-
-        // POST: Rooms/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
-        {
-            Rooms rooms = db.Rooms.Find(id);
-            db.Rooms.Remove(rooms);
-            db.SaveChanges();
-            return RedirectToAction("Index");
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing)
-            {
-                db.Dispose();
-            }
-            base.Dispose(disposing);
-        }
-    }
-}
