+++ Software_Engineering_Project/Software_Engineering_Project.csproj
@@ -161,7 +161,8 @@
     <Compile Include="Models\Descriptors.cs" />
     <Compile Include="Models\Rooms.cs" />
     <Compile Include="Models\RoomsContext.cs" />
-    <Compile Include="Models\RoomsController.cs" />
+    <Compile Include="Controllers\RoomsController.cs" />
+    <Compile Include="Models\AdminDashModel.cs" />
     <Compile Include="Models\Users.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
